Issue 1599, 1601: Get date popups to properly dynamically updated via a work-around.
This is an unideal work-around to get things moving forward.
Use a separate variable to apply a "fixed" value to not deal with datepicker problems and then handle that value on save.

Move the code into the submission so that update events can be processed.
This also helps reduce redundant code.

Add aback `stringtodate` which was removed for testing purposes and the change was accidentally committed in the previous commit.