Issue 1599, 1601: Relocate common code and fix additional case where popup date is wrong.
Move the work-around out of the field profile directive and out of the submission info directive.
Use a single function on the submission model as the new location for both cases.

Even after applying the filter the date popup occasionally gets it wrong.
This time the corrected date is not consistently being corrected due to the JavaScript `Date()` behavior.
Once the offset is applied use that offset date to manually set the date, fixing any problems.

Reflect the recent changes in the unit test mocks by adding stub functions.

Remove no longer used datetime HTML file.