Issue 1709: Add type checks to prevent errors when handling invalid dates.
The Javascript date function is very sensitive to values.
The regex match function does not always return an array.

While it would be preferred that a Javascript regex match returns an empty array it does not.
Handle the error case for when the regex match fails (such as on an empty string).
This fixes the display problem on the admin list view.

The `date()` function may return the string 'invalid date' or something similar.
This then causes the `date.toISOString()` to fail.
Oddly enough, the `isNaN()` will succeed for when `date()` is a valid response (which is an Object).
Using the `isNaN()` check fixes the display problem with using the submission view confirmation page (step 5 of the submission view for "Confirm & Submit").

Re-organize and add a NULL check to `$scope.displaySubmissionProperty()` just in case.

Clean up some white space in related areas of code.