Issue 1582: Reset list view pagination and fix exposed pagination related problems.
Change the behavior to reset the pagination when a filter is applied.
This prevents the bug where the list is empty and the table is unusable when a filter is applied.

Making this change exposed several related problems and bugs that have been deemed necessary to be fixed as part of this issue.

A new `resetPagination()` function is created that performs the reset task.
This required replacing duplicated code that also performs the pagination.
This change exposed a bug where the `update()` function is always calling the database list view load even when unnecessary.

This additional list load request in `update()` ends up altering the pagination and must be fixed to get this new behavior to work consistently.
Preventing it from reloading the list when not needed also improves performance by removing an extra unnecessary request in certain places.
Break apart the `update()` function, adding an additional function `processUpdate()`.
This function performs the tasks needed.
Breaking this up is required because there is an additional request to `ManagerSubmissionListColumnRepo.submissionListPageSize()` that alters the page size.
The `ManagerSubmissionListColumnRepo.submissionListPageSize()` effectively reset the selected page count, causing a regression.
The new behavior is to only call `ManagerSubmissionListColumnRepo.submissionListPageSize()` when there is no local session data.
This design change maintains pre-existing behavior with the page count.
This is another likely performance improvement where the page count is no longer fetched when not necessary.

Despite the improvements, I still found a case where the page list is empty and the table is broken.
An example would be data changes since pagination is set.
It turns out that all of the data needed to determine if the selected pagination is correct is returned on the list response.
The problem is that the list response is already loaded so even if the pagination data were to be corrected there is no row data to display.
This is solved by making an extra request with the pagination data corrected using data from the first request.
Recursion is completely undesirable and so this extra request is performed exactly once and then only in the specific situation where the first request is incorrect.
Doing this required breaking apart the `queryGetData()` and adding a new parameter for forcing the page number.

Removing a single row from the list view should not reset the pagination, unlike the other filters.
When a filter that removes a single row is restored then the pagination must also not be reset.
This preserves existing expected behavior.
This exposed another bug is observed where when the pagination is restored page index may be incorrect.
This is because the database stores the page numbers starting at 0 while the UI table is storing page numbers starting at 1.
There are some cases where 1 is incorrectly added or not added.
Fixing this ended up being necessary as part of solving the original issue requested to be fixed.
The solution here is that when the value comes from the database, increment it by one.
The value is preserved, unchanged, in all other cases.

When removing a single row and that row is the last row for the page then the pagination must be updated.
Change the current page number to the previous page in the case where the current page is not the first page (page index 0 as per DB or page index 1 as per UI).