Issue 1731: Improve handling when attempting to delete filters.
The failure may be happening due to problems with finding the filter in the saved filters and it not being there.
Add sanity checks and then only attempt the save when the filter actually exists in the saved filter for the current user.
Otherwise, return a failure.

Rename the `removeFilter` into `removeSaveFilter` to be less confusing and more consistent with similar functions in the UI.

This depends on commit 0820fce44bf02cb8c1d8a7e47ecb27bb180654d0 and is based off of the branch for the PR #1736 that contains the referenced commit.