Issue 1667: Reduce cost of the COUNT query for the submission list dynamic query.
This helps with the performance problem.

The COUNT query only needs the ID and does not need group by or any of the columns.
The sub-queries with their joins are causing problems.
If any WHERE condition is needed, use a single join query rather than a sub-query.
This is safe because ORDER BY is not needed in the count query.

The non-COUNT query should also not add any sub-queries unless those sub-queries have a WHERE condition or a ORDER BY clause association with it.
This does not address the problems with the non-COUNT query.

This does not solve the problem at hand, instead it is progress towards the solution.

This adds additional debug logging to show the timings between the COUNT, non-COUNT, and other queries.