Issue 1667: The user settings active filter is EAGER.
This is believed to be the primary cause of the performance problem observed when a user has an active filter.
The performance cost can be seen ot only on the submission list but on just about any other page so long as a filter was made active of the submission list at one point in time.