Issue 1667: The non-count query should not join the sub-queries if not necessary.
The query is designed to only return the submission ids.
The individual columns are added in the SELECT line because, for whatever reason, SQL requires that anything in the ORDER BY to be in the column.

There are, therefore, two cases when the sub-query must be joined:
1. When there is a condition for the WHERE clause (the filters).
2. When the ORDER BY is in use for what would be represented in that sub-query.

In all other cases, the non-COUNT SELECT query can be made similar to the COUNT query such that only the required sub-queries are joined.

Another query is made later on in the code that might pull in the required columns.
That query is passed the submission ids that come from this query.