Issue 1706: Remove the auto-focus from the submission.
There are two focus directives that each play a part in the auto-focus.
Removing them both from the submission form solves the problem.
Completely remove the directives and tests because they are not used anywhere else.

The auto-focus directive is only used on the submission view.
The field profile directive may be used in more places.
The field profile directive very clearly is causing the focus problem.