Issue 1764: Use aria-details for the license agreement and proquest agreement.
This should improve accessibility by better associating the agreement checkbox with the agreements.
The "I agree" check boxes are not described by the agreement.
Instead, the "I agree" checkboxes is instead associated with the agreement.
The `aria-describedby` is not as semantically ideal in this case as is `aria-details`.

The `aria-describedby` should likely be avoided also because of the size of the agreements.

Consider this: "Unlike aria-describedby, elements referenced by aria-details are not used in accessible descriptions and are not turned into a plain string when presented to assistive technology users. If the associated content is not too long and flattening the contents of the referenced element to a simple string of text wouldn't cause loss of information, consider using aria-describedby instead. That said, it is valid for an element to have both aria-details and a description specified with either aria-describedby or aria-description."
Consider this: "Elements referenced by aria-details are intended to contain more information than would normally be provided via aria-describedby."
source: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-details