Issue 1739: Then user delete is not properly deleting the repository.
Each individual saved filter must be deleted not the active filter.
The active filter can be unset as it either points to a "saved" filter or a persisted filter.

In the case of a persisted filter, the filter might get auto-deleted.
This is neither tested nor confirmed and investigation is needed.

Use `super.delete()` to trigger all of the appropriate normal JPA delete functionality.