Issue 1739: Add transactional to unit tests and disable out of scope problems.
The `@Transactional` is now required to ensure `lazy` loaded data is not a problem.
This then exposes additional problems and also requires additional special handling.

Most tests should be run with `Propagation.REQUIRES_NEW`.
Some tests cannot be run in a transaction, so use `propagation = Propagation.NEVER` for those.

The `setup()` and `cleanup()` methods should no longer be provided by the abstract for `AbstractIntegrationTest`.
The `cleanup()` are entirely removed because they are problematic and the transaction eliminates their need.

The `webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT` must be set to run individual tests using something like `mvn clean test -Dtest=ActionLogTest`.

Some tests have to be disabled due to problems.
Comments are added regarding their problems, such as the method named for testing the delete only calls a create and does not call a delete.

Disable the stub tests.
This might not be ideal as it says "skipped" but these are stub tests that test nothing anyway.

The `UserTest` is the most problematic when it comes to transactions.
The need for using `propagation = Propagation.NESTED` has been determined.
The `propagation = Propagation.REQUIRED` has been added but may not be required.
The assertion test is changed in order to better function with the transaction.