Issue 1780: Only create filters for appropriate roles.
When the `saveAndAddSettings()` is called, such as on user creation, make sure the active filter is created only for the roles `ROLE_ADMIN`, `ROLE_MANAGER`, and `ROLE_REVIEWER`.

When updating the user, check again the role.
If the role is one of the above, then check to see if the active filter is NULL.
Checking for NULL here is better in that it avoids additional database queries to get the actual previous role.
The problem being solved is not whether the user used to be some other role.
The problem being is that if the user does not have an active filter then there are problems.

The active filter being NULL is only a problem for the above roles.