Issue 1780: Replace userRepo.save() calls with userRepo.update() calls.
The role settings are not being properly setup.
This is because the `userRepo.save()` is being called and the `initializeActiveFilter()` is not being called on `userRepo.save()`.
The `userRepo.save()` is not easily overridable.

This is the quickest solution to the problem at hand given the short time frame that I have available to address this situation.

I also noticed that the `updateSetting` end point (`/settings/update`) is calling `userRepo.save()` after calling `userRepo.update()`.
Remove the redundant `userRepo.save()`.