Issue 1756: Increase batch size in ControlledVocabulary and use SUBSELECT in VocabularyWord.
The `ControlledVocabulary` is switched to `LAZY` and sets the batch size to 1000 for the dictionary property.
The increased batch size should reduce the amount of queries for larger data sets.
The `LAZY` is set to not immediately load the dictionary.

A major performance problem is within the `VocabularyWord` Entity where all contacts are loaded.
There may be as many as 20,000 contacts.
Loading every single one as a fully populated Entity Object (and then one at a time) is very costly.
Reduce the impact of this by using a `SUBSELECT`.

Furthermore, use `LAZY` in more properties on the `VocabularyWord` Entity.

This approach does not solve the problem.
Instead, it reduces the impact.
For a set of maybe 17,000 contacts, the loading of a single organization may take up to 22 seconds.
After this change it may take up to 4 seconds.
This is still too long but is far better.

The problem is in how the Entity is being loaded.
Other strategies need to be developed (such as using Projection).
Such strategies are not performed within this commit.