Issue 1756: Use Projection for specific Submission usage.
This is affecting the **Choose Your Organization** page.

Too much unnecessary data is being loaded off of the specific Submission end point `/all-by-user`.
This end point is only used for the main page that then leads to the **Choose Your Organization** page.

This reduces the loaded data to only the minimum needed on the Student Submission List View.
The observed difference in performance is ~800 milliseconds prior to this change and ~200 milliseconds after this change (for vocabulary words with ~17,000 contacts).
Also note that Without the code from commit a2e872a6fb37a8bf1ede8cbb8210455a62ebff27, the prior performance is ~22 seconds (for vocabulary words with ~17,000 contacts).