Issue 1756: Assignee must have field values attached.
This actually only needs to be values for two predicates:
1) `dc.title`
2) `_doctype_primary`

The problem is the current design does not make this easy.
Rather than taking a more complicated approach, suffer the performance cost and load all field values for the submissions.
In the particular data set that I am using (a copy of production data), the request goes from ~200ms to ~300ms with this change.
I consider this an acceptable loss for the time being.
Future designs should be applied to avoid this and load only the two values for the specified predicates.