Issue 1756: CLI Generation broke after switching to LAZY initialization.
Must now add `@Transactional`.

I moved the relevant code blocks into their own functions and added `@Transactional`.
This turned out not to work.
The `@Transactional` must be on the `public void run()` function.

I left the separate functions as-is because it is cleaner code.

The expansive is taking a bit long sometimes so I lowered the action log max and provided a second parameter to make it configurable.
```
  generate 5 20
```
Would Generate 5 submissions with a max of 20 action logs randomly generated when in expansive mode.

Improve some messages.

Simplify some of the arguments, moving them out of the switch statement (making the coder slightly cleaner).