Issue 1819: Improve reporting of the Sword error messages.
This is based off of the commit 07a4f3ce700532b7f9ee1a0565950e967cb43f97.

The error message returned is misleading.
Change the error message to report the actual reason for failure so that better debugging can be performed.
Add new Response State exceptions (this approach taken is a bit tedious but it works).

This takes a conservative approach and re-uses existing error handling.
This existing error handling still alters the message but is more accurate than the previous behavior.

Move the ServiceDocument declaration outside of the try block so that it can be used to determine if the failure is due to being unable to fetch the service document for any other reason.
I don't know why the original error message is not displayed and so I am maintaining the existing behavior.

For all other error messages, just return HTTP 400.