Issue 1795: Redesign filter broadcasting logic.
This is a work in progress commit.

The problem is that the model listener is on a per individual filter basis.
The entire set for both saved filters and active filters for a given user needs to be handled as a whole.

Remove or disable the model listener, replace the old end points, and add new broadcasts and listeners that operate on a set basis rather than on individual basis.