Issue 1795: Prevent double-loading of query list and appropriately reset pagination.
The changes come in from the broadcast.
The recipient of the broadcast doesn't necessarily know what the broadcast is for.
Provide additional details about the broadcast so that the list can be more appropriately queried and built.
A new enumeration is provided to facilitate this.
Only the most basic parts of the enumeration are created.

Prevent the double-loading of the query by removing the direct calls to refresh the query.
Instead, let the broadcast listener handle the calling of the query.