Issue 1795: Public saved filters should update for all users who can see it.
Only two actions are needed here:
1. `SAVE`
2. `REMOVE`

All users who can see a public filter should have their saved filter list up date when a public filter is added or removed.

The `SAVE` will only refresh the list if the user is different from the user who created the filter.
The user who created the filter will already be up to date due to the other broadcasts.

The `SAVE` also needs to handle the case where the owning user changes the filter from public to private or vice versa.

The `REMOVE` needs special handling because the object is removed.
Send only the ID.
Future iterations may want to also pass the user ID to allow for detecting if the current user is the one who deleted the filter.
Currently, passing both the Filter ID and the User ID results in overlap because they both get serialized into a key named `Long` and as a result one overwrites the other.