Hotfix: Making changes after selecting a saved filter alters the saved filter.
The filter should not be altered.
This hot fix takes a quick approach of just always assigning the "persisted" filter as the active filter.

When this filter gets modified, the "persisted" filter now is what gets modified rather than the saved filter.

For future notes and investigations consider that this solution also suggests that the "active filter" concept can be thrown away and the "persisted filter" could always be the active filter.