Issue 1817: Move controller-time method creation off of model and onto the controller.
The `$scope.modalData.testDepositLocation()`, `$scope.modalData.isTestDepositing()`, and `$scope.modalData.isTestable()` methods are being manually created on the `$scope.modalData` for every `$scope.resetDepositLocation()` call.

These methods should either be on the modal or on the controller.
These methods include controller specific code and are now moved to the controller.
These methods are now not constantly being re-created on every `$scope.resetDepositLocation()` call.

The unit tests are now fixed and the `$scope.isTestable()` is changed to properly return booleans by utilising the double-not notation: `!!`.

This problem was first observed by wwelling.
Co-authored-by: William Welling <wwelling@library.tamu.edu>