Issue 1817: Relocate the testConnection() function off of the Model and onto the Repository.
The problem is that the `testConnection()` is on the model.
The model is not saved when the `testConnection()` is called.
The model uses its own data, which is only the saved data.

Relocate the `testConnection()` off of the model and onto the repository.

The models really shouldn't be performing network actions anyway.
The repositories are intended for this purpose.