Issue 1827: The wrong URLs are being used to deposit.
The current behavior is to set the URL to `servicedocument` to perform the testing.
This ends up being used by both the `getCollections()` and `deposit()` end points.

If the Repository URL is set to say `http://example.com/sword/servicedocument`, then this works for the `getCollections()` and breaks for the `deposit()` end points.
If the Repository URL is set to say `http://example.com/sword/deposit`, then this breaks  for the `getCollections()` and works for the `deposit()` end points.

Both need to work.
Change the Repository URL configuration design to instead except a Repository URL of `http://example.com/sword/`.
Now, when `getCollections()` end point is called, it appends the `/servicedocument` to that URL.
Now, when `deposit()` end point is called, it appends the `/deposit` to that URL.

This could be simplified further to append to not have the `/sword` and then append that as needed.
However, this change is considered out of scope and is left for future discussion.