Issue 1827: Add configuration option to toggle swordv1 repository URL behavior.
Add a configuration file witht he configuration `vireo.depositor.swordv1.single-url` to designate how to handle the URL.
This allows for backwards compatibility in case the new behavior is a problem in some system.

The `SWORDv1Depositor` itself is not detectable by Spring and so `@Autowired` does not work.
Fix this by adding an `@Bean` in the configuration and calling that loader instead of `new SWORDv1Depositor()` in the system data loader.