Issue 1807: Deposit ID and Notes list problems and escape SQL improvements.
The Deposit ID and the Notes are both not being filtered and displayed properly on the Submission List admin view page.

The `depositURL` must be set to the JsonView of `Views.SubmissionList.class` to allow for the data to be transferred.

The columns must be selected in the SQL query in the "alias" part of the query.
Change the logic to handle this.
Fix an SQL escape security concern while working on this.

The System default for the Submission List columns is missing the appropriate `valuePath`.
The `valuePath` must be set to exactly what the Submission model returns (case-sensitive), which happens to be `depositURL`.