Issue 1827: The repository URL is misleading and gets stripped.
Partially restore the previous behavior of using the `/servicedocument`.
It turns out that while the repository URL gets used, it gets stripped to just the host name and port number during deposit.
The swordv1 library ends up adding the deposit location from the service document manifest.

Using the Repository URL in the error logs is misleading and confusing.
Change the behavior to better describe what the given URL is on error.
The sword handles the actual deposit location internally and is not as readily available for use in the error log.