Issue 1835: Fix Non-Chairing Committee Contact E-mails to show up on forms.
Previous sprints improved performance by adding `!JsonView` and causing this regression.

The Vocabulary Word needs to include the contacts e-mails for the Field Predicate and Field Profile.
Add `@JsonView(Views.Partial.class)` to the contacts in Vocabulary Word.

This, however, is not enough.
Both the Field Predicate and the Field Profile need to be set to at least `Views.Partial.class`.

When this is done, the contact e-mail will not transfer to the UI as expected and required.
The JsonView changes from previous sprints changed the