Don't use both `@Spy` and `@InjectMocks` on same property and make simpMessagingTemplate and env protected rather than private.
This could be causing a race conditionition during the loading of the `@Spy`.
The `@InjectMocks` likely does everything needed and the `@Spy` is probably not needed in this case.

Both simpMessagingTemplate and env are mocked.
An implementing class needs to utilize the mock then the implementing class needs access to the property