Issue 1840: Delete Field Profile by ID.
Change the behavior to accept an ID.
Add error checking and returning ERROR as appropriate.

To avoid changing the API too much, change the request body to be the Field Profile ID rather than the entire Field Profile (as opposed to sending the id via the URL as a path parameter).