Issue 1842: Submission load time is affected by fix for issue 1842.
The Issue 1842 (via commit f4494041060e5451df04e37aad0093eb8886f71e) introduced the `ControlledVocabularyRepo` onto the Field Profile directive.
The result of this is that when the Student Submission view loads the submission for editing, the entire Controlled Vocabulary ends up being loaded.
This has a notable performance hit.

Set the mapping to lazy to prevent this loading from happening and thereby improving performance.