Issue 1802: Manuscript File name is not being loaded.
On the Submission History view, the manuscript file name needs to be manually loaded using the submission call.
Do the fetch manually, updating the value as needed.
Change the behavior to have the default in the parameter rather than via an `||`.

A locking mechanism is in use to prevent problems with the digest cycles, ensuring only a single load.
The locking is implemented on a per field value id basis.

Simplify `enrichDocumentTypeFieldValue()` design to use a single check on `fieldValue.value`.