Issue 1858: Fix WYSIWYG in lock unlock fields, reverting previous approach.
Originally, the Locking Text Area fields were not working at some point in time.
This got fixed by disablying the wysiwyg attribute in this PR:
- https://github.com/TexasDigitalLibrary/Vireo/pull/1792

(For this issue: https://github.com/TexasDigitalLibrary/Vireo/issues/1776)

The lack of a `wysiwyg` attribute suggested that wysiwyg is supposed to be false.

That code was approved and merged in previous sprints.

The same bug was discovered a second time in different places and recorded here:
- https://github.com/TexasDigitalLibrary/Vireo/issues/1858

The solution chosen is to do the same exact approach as the previously approved code:
- https://github.com/TexasDigitalLibrary/Vireo/pull/1859

The problem is not the "=== 'wysiwyg'" and "!== 'wysiwyg'" checks.

Instead, the WYSIWYG data is not being loaded by the webpack.
This is essentially a regression from when the project switched to use webpack.